requirements.txt
fastapi
langchain
langchain-google-genai
beautifulsoup4
requests
python-dotenv
youtube-transcript-api
uvicorn